Update SampleActivity.kt to address issue #55
Because of the time it takes to initialize and create the player after the call to `dm_player_web_view.load({video_id and/or params})`, a check needs to be made for edge cases when users leave the activity while the player view is still loading. This can easily be achieved by checking if the player has any buffered time. 

It will if fully loaded and thus the call to `dm_player_web_view.onPause()` will be successful. If the buffered time is 0, then the player hasn't fully initialized/loaded yet and runs the risk of finishing in the background if the user exits during that time. Loading the player view with a null value prevents this behavior and reloading with desired video on user return results in a seamless user experience that conforms with the Android platform standards.